split frontend diagnostics fixtures
